# lib/model: Microoptimization of unifySubs and blockDiff