--- a/lib/model/model.go
+++ b/lib/model/model.go
@@ -2771,48 +2771,29 @@ func readOffsetIntoBuf(fs fs.Filesystem, file string, offset int64, buf []byte) 
 // The exists function is expected to return true for all known paths
 // (excluding "" and ".")
 func unifySubs(dirs []string, exists func(dir string) bool) []string {
-	subs := trimUntilParentKnown(dirs, exists)
-	sort.Strings(subs)
-	return simplifySortedPaths(subs)
-}
-
-func trimUntilParentKnown(dirs []string, exists func(dir string) bool) []string {
-	var subs []string
-	for _, sub := range dirs {
-		for sub != "" && !fs.IsInternal(sub) {
-			sub = filepath.Clean(sub)
-			parent := filepath.Dir(sub)
-			if parent == "." || exists(parent) {
-				break
-			}
-			sub = parent
-			if sub == "." || sub == string(filepath.Separator) {
-				// Shortcut. We are going to scan the full folder, so we can
-				// just return an empty list of subs at this point.
-				return nil
-			}
-		}
-		if sub == "" {
-			return nil
-		}
-		subs = append(subs, sub)
+	if len(dirs) == 0 {
+		return nil
 	}
-	return subs
-}
-
-func simplifySortedPaths(subs []string) []string {
-	var cleaned []string
-next:
-	for _, sub := range subs {
-		for _, existing := range cleaned {
-			if sub == existing || strings.HasPrefix(sub, existing+string(fs.PathSeparator)) {
-				continue next
-			}
-		}
-		cleaned = append(cleaned, sub)
+	sort.Strings(dirs)
+	if dirs[0] == "" || dirs[0] == "." || dirs[0] == string(fs.PathSeparator) {
+		return nil
 	}
-
-	return cleaned
+	unified := dirs[:0]
+	prev := "./" // Anything that can't be parent of a clean path
+	for _, dir := range dirs {
+		dir = filepath.Clean(dir)
+		if dir == prev || strings.HasPrefix(dir, prev+string(fs.PathSeparator)) {
+			continue
+		}
+		parent := filepath.Dir(dir)
+		for !(exists(parent) || parent == "." || parent == string(fs.PathSeparator)) {
+			dir = parent
+			parent = filepath.Dir(dir)
+		}
+		unified = append(unified, dir)
+		prev = dir
+	}
+	return unified
 }
 
 // makeForgetUpdate takes an index update and constructs a download progress update