Recently unnecessary allocs came up in some comments and I noticed some in the code. This "optimization" is cosmetic and the benchmarks are definitely over the top, but as I have already written them out, why not keep it. Also I personally consider the rewritten `unifySubs` simpler.

```
BenchmarkUnifySubs-4      200000              7947 ns/op            1008 B/op        36 allocs/op
BenchmarkUnifySubs-4      500000              3143 ns/op             256 B/op         8 allocs/op

BenchmarkDiff-4           300000              5334 ns/op            1920 B/op        27 allocs/op
BenchmarkDiff-4           300000              5013 ns/op            2592 B/op        24 allocs/op
```