@st-review 2018-01-14
@AudriusButkevicius: Noted! Need another LGTM or explicit merge command.