@st-review 2018-01-14
:ok_hand: Merged as bd63fd73b149ca41122f901f9d9fa9655942b42c. Thanks, @imsodin!